fix compile

+++ NaiveStudio/Script/FunctionManager.h
@@ -12,7 +12,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface FunctionManager : NSObject
 
-+ (NSString *)statementOfGetObjectWithAddress:(NSUInteger)address;
++ (NSString *)statementOfGetObjectWithAddress:(unsigned long long)address;
 
 @end
